pyproject.toml
[tool.poetry]
name = "comicpy"
version = "0.1.0"
description = ""
authors = ["kurotom <55354389+kurotom@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
pypdf = "^4.0.1"
pillow = "<=10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
